requirements.txt
tweepy==3.5.0
nltk
pyyaml
